// src/api-client/models/Category.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Budget Manager API
 * Personal finance management API for tracking income, expenses, accounts, and budgets.
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { CategoryType } from './CategoryType';
import {
    CategoryTypeFromJSON,
    CategoryTypeFromJSONTyped,
    CategoryTypeToJSON,
    CategoryTypeToJSONTyped,
} from './CategoryType';

/**
 * 
 * @export
 * @interface Category
 */
export interface Category {
    /**
     * 
     * @type {string}
     * @memberof Category
     */
    id?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Category
     */
    userId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Category
     */
    name?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Category
     */
    image?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Category
     */
    color?: string | null;
    /**
     * 
     * @type {CategoryType}
     * @memberof Category
     */
    categoryType?: CategoryType;
    /**
     * 
     * @type {Array<string>}
     * @memberof Category
     */
    subcategories?: Array<string> | null;
    /**
     * 
     * @type {Date}
     * @memberof Category
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof Category
     */
    updatedAt?: Date;
}



/**
 * Check if a given object implements the Category interface.
 */
export function instanceOfCategory(value: object): value is Category {
    return true;
}

export function CategoryFromJSON(json: any): Category {
    return CategoryFromJSONTyped(json, false);
}

export function CategoryFromJSONTyped(json: any, ignoreDiscriminator: boolean): Category {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'userId': json['userId'] == null ? undefined : json['userId'],
        'name': json['name'] == null ? undefined : json['name'],
        'image': json['image'] == null ? undefined : json['image'],
        'color': json['color'] == null ? undefined : json['color'],
        'categoryType': json['categoryType'] == null ? undefined : CategoryTypeFromJSON(json['categoryType']),
        'subcategories': json['subcategories'] == null ? undefined : json['subcategories'],
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
        'updatedAt': json['updatedAt'] == null ? undefined : (new Date(json['updatedAt'])),
    };
}

export function CategoryToJSON(json: any): Category {
    return CategoryToJSONTyped(json, false);
}

export function CategoryToJSONTyped(value?: Category | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'userId': value['userId'],
        'name': value['name'],
        'image': value['image'],
        'color': value['color'],
        'categoryType': CategoryTypeToJSON(value['categoryType']),
        'subcategories': value['subcategories'],
        'createdAt': value['createdAt'] == null ? value['createdAt'] : value['createdAt'].toISOString(),
        'updatedAt': value['updatedAt'] == null ? value['updatedAt'] : value['updatedAt'].toISOString(),
    };
}