// src/api-client/models/Transaction.ts
export interface Transaction {
    /**
     * 
     * @type {string}
     * @memberof Transaction
     */
    id?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Transaction
     */
    userId?: string | null;
    /**
     * 
     * @type {TransactionType}
     * @memberof Transaction
     */
    transactionType?: TransactionType;
    /**
     * 
     * @type {number}
     * @memberof Transaction
     */
    amount?: number;
    /**
     * 
     * @type {Date}
     * @memberof Transaction
     */
    date?: Date;
    /**
     * 
     * @type {Date}
     * @memberof Transaction
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof Transaction
     */
    updatedAt?: Date;
    /**
     * 
     * @type {string}
     * @memberof Transaction
     */
    accountId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Transaction
     */
    accountName?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Transaction
     */
    toAccountId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Transaction
     */
    toAccountName?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Transaction
     */
    fromAccountId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Transaction
     */
    fromAccountName?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Transaction
     */
    categoryId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Transaction
     */
    categoryName?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Transaction
     */
    categoryImage?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Transaction
     */
    categoryColor?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Transaction
     */
    subcategory?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Transaction
     */
    notes?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof Transaction
     */
    isApplied?: boolean;
    /**
     * 
     * @type {number}
     * @memberof Transaction
     */
    appliedAmount?: number | null;
    /**
     * 
     * @type {boolean}
     * @memberof Transaction
     */
    isMonthly?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof Transaction
     */
    isRecurring?: boolean;
    /**
     * 
     * @type {string}
     * @memberof Transaction
     */
    monthlyKey?: string | null;
    /**
     * 
     * @type {number}
     * @memberof Transaction
     */
    recurringTimes?: number | null;
    /**
     * 
     * @type {string}
     * @memberof Transaction
     */
    recurringType?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Transaction
     */
    savingKey?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Transaction
     */
    transferId?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof Transaction
     */
    removeFromSpendingPlan?: boolean;
    /**
     * 
     * @type {string}
     * @memberof Transaction
     */
    yearMonth?: string | null;
    /**
     * 
     * @type {number}
     * @memberof Transaction
     */
    year?: number;
    /**
     * 
     * @type {number}
     * @memberof Transaction
     */
    month?: number;
    /**
     * 
     * @type {number}
     * @memberof Transaction
     */
    day?: number;
    /**
     * 
     * @type {{ [key: string]: any; }}
     * @memberof Transaction
     */
    metadata?: { [key: string]: any; } | null;
    /**
     * 
     * @type {string}
     * @memberof Transaction
     */
    type?: string | null;
}